Auto-Shoot Code Proof of Concept
Added variables for shooter masterposition and velocity. Upped RPMs on test fire option. Most importantly, added code for test fire mode that automatically runs conveyer when shooter is triggered - overrides normal flow of conveyer, but is itself still overrridden if conveyer is in manual mode. Not yet fully implemented to rest of shooter (may leave for rest of team to implement)